Reimplementation of run, parse, typecheck and compile (#692)
These functions are updated in their types: instead of being called as a
chain (compile calls typecheck, which calls parse), they are now
standalone and should be called in the right order in run. This enables
a much cleaner pipeline. In the future more functions in the pipeline
should be ported to use this method. Feedback will provide a unified
mechanism for keeping track of data and messages.

For now the functions are changed minimally to support Feedback.