Implemented `msg` in Feedback as an Applicative.
By relying on an applicative, `app` in this case, we can insert single
messages without relying on any knowledge of the specific applicative.
This will come in handy for writing an instance of MonadFail:
```
fail :: String -> Feedback app msg a
```
Here we will need to translate a `String` into a `msg`, that is
contained in applicative `app`. Without the applicative, we would need
to translate a `String` into a datatype that is hidden inside another
type (the `msg` type before the changes).